backend/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
psycopg==3.1.9
psycopg-binary==3.1.9
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3